+++ src/main/java/co/edu/uceva/buildcheck/modules/proyectos/service/ProyectoServiceImpl.java
@@ -2,6 +2,7 @@ package co.edu.uceva.buildcheck.modules.proyectos.service;
 
 import co.edu.uceva.buildcheck.modules.proyectos.model.Proyecto;
 import co.edu.uceva.buildcheck.modules.proyectos.repository.IProyectoRepository;
+import co.edu.uceva.buildcheck.modules.movimientos.repository.MovimientoRepository;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,10 +11,12 @@ import java.util.Optional;
 
 @Service
 public class ProyectoServiceImpl implements IProyectoService{
-    IProyectoRepository proyectoRepository;
+    private final IProyectoRepository proyectoRepository;
+    private final MovimientoRepository movimientoRepository;
 
-    public ProyectoServiceImpl(IProyectoRepository proyectoRepository){
+    public ProyectoServiceImpl(IProyectoRepository proyectoRepository, MovimientoRepository movimientoRepository){
         this.proyectoRepository = proyectoRepository;
+        this.movimientoRepository = movimientoRepository;
     }
 
     @Override
@@ -25,7 +28,11 @@ public class ProyectoServiceImpl implements IProyectoService{
     @Override
     @Transactional
     public void delete(Proyecto proyecto){
-        proyectoRepository.delete(proyecto);
+        if (!movimientoRepository.existsByProyecto(proyecto)) {
+            proyectoRepository.delete(proyecto);   
+        }else{
+            throw new IllegalStateException("No se puede eliminar el proyecto porque tiene movimientos asociados");
+        }
     }
 
     @Override
@@ -45,4 +52,5 @@ public class ProyectoServiceImpl implements IProyectoService{
     public List<Proyecto> findAll(){
         return proyectoRepository.findAll();
     }
+
 }
